# Conflicts:
#	src/main/java/org/team1/keyduck/auction/dto/response/AuctionReadAllResponseDto.java
#	src/main/java/org/team1/keyduck/auction/dto/response/AuctionReadResponseDto.java

+++ src/main/java/org/team1/keyduck/member/service/MemberService.java
@@ -4,9 +4,13 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.team1.keyduck.auction.entity.AuctionStatus;
+import org.team1.keyduck.auction.repository.AuctionRepository;
+import org.team1.keyduck.auth.service.JwtBlacklistService;
 import org.team1.keyduck.common.exception.DataInvalidException;
 import org.team1.keyduck.common.exception.DataNotFoundException;
 import org.team1.keyduck.common.exception.ErrorCode;
+import org.team1.keyduck.common.exception.OperationNotAllowedException;
 import org.team1.keyduck.member.dto.request.MemberUpdatePasswordRequestDto;
 import org.team1.keyduck.member.dto.request.MemberUpdateRequestDto;
 import org.team1.keyduck.member.dto.response.MemberReadResponseDto;
@@ -21,6 +25,8 @@ public class MemberService {
 
     private final MemberRepository memberRepository;
     private final PasswordEncoder passwordEncoder;
+    private final AuctionRepository auctionRepository;
+    private final JwtBlacklistService jwtBlacklistService;
     private final PaymentDepositRepository paymentDepositRepository;
 
     @Transactional
@@ -50,9 +56,20 @@ public class MemberService {
     }
 
     @Transactional
-    public void deleteMember(Long id) {
-        Member member = memberRepository.findById(id).orElseThrow(() -> new DataNotFoundException(
-                ErrorCode.NOT_FOUND_MEMBER, "멤버"));
+    public void deleteMember(Long id, String token) {
+        Member member = memberRepository.findByIdAndIsDeleted(id, false);
+
+        if (member == null) {
+            throw new DataNotFoundException(ErrorCode.NOT_FOUND_MEMBER, "멤버");
+        }
+
+        //현재 진행중인 경매가 있으면 탈퇴 불가능
+        if (auctionRepository.existsByKeyboard_Member_IdAndAuctionStatus(id,
+                AuctionStatus.IN_PROGRESS)) {
+            throw new OperationNotAllowedException(ErrorCode.DELETE_FAIL_AUCTION_IN_PROGRESS, null);
+        }
+
+        jwtBlacklistService.addToBlacklist(token);
 
         member.deleteMember();
     }
